requirements.txt
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
faiss-cpu==1.11.0
filelock==3.18.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
griffe==1.7.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.42
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.6.0
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.37.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.5
openai==1.75.0
openai-agents==0.0.9
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
prov==2.0.1
puremagic==1.29
pyarrow==20.0.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pydot==4.0.0
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.1
streamlit==1.45.0
tavily-python==0.7.2
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
yarl==1.20.0
zstandard==0.23.0
